+++ CHANGELOG.md
@@ -1,8 +1,9 @@
 # Change Log
 
 ## [Unreleased]
-### Fixes
+### Added
 - Added support to manually set bulk size and enable/disable bulk dispatch
+- `loadConfigurations` and `loadConfigurationsFromJson` functions
 
 ## [1.0.0-beta.1]
 ### Changes